use switch/case to parse commands
seems cleaner and easier to maintain.
reuse code with fallthrough.
add a new case for the new command.
all there is to do, is to set the
REPLY var to the text you want to send.
if no message should be sent then
add continue, if iii should stop
then add break.