iii: revisit nick coloring and cleanup braces
Nicks now take colors based on their first 3 letters
and the nick length.

The nick is first sanitized; trailing numbers 1 and 2
or an underscore are trimmed, as they are though of
as indicators of unregistered, failed to identify, or
duplicate due to ghost users, nicks.

Additionally, some colors are excluded as they seem
too dark or bright to be useable.

Finally, certain "special" nicks are colored dark,
both their nicks and messages; this applies to bots
and server messages.